# feat(perf): speed up TS AVM core simulator