… noir-projects/noir-contracts/contracts/avm_test_contract/src/main.nr
… yarn-project/simulator/src/avm/avm_memory_types.test.ts
… yarn-project/simulator/src/avm/opcodes/conversion.ts
… yarn-project/simulator/src/avm/opcodes/environment_getters.ts
… yarn-project/simulator/src/avm/opcodes/external_calls.ts
… yarn-project/simulator/src/avm/opcodes/hashing.ts
… yarn-project/simulator/src/avm/opcodes/memory.ts
… yarn-project/simulator/src/avm/opcodes/misc.ts
… yarn-project/simulator/src/avm/opcodes/multi_scalar_mul.ts
… yarn-project/simulator/src/avm/opcodes/storage.ts