## Summary
This PR cleans up and optimizes the core AVM instruction execution. This does NOT include any optimizations to merkle/db-access opcodes, but instead the optimizations are meant to speed up apps that execute a huge number of small opcodes.

## Testing
To test this PR, I ran sha256 of 2048 bytes, which executes 163,760 instructions (primarily MOV, ADD, LT, EQ). These instructions do very little other than charge gas, load/store memory, check memory tags, and logger.trace().


## 66% speed improvement 
This is of course for a test of *only the core simulator* that runs a massive loop of *160k _simple_ instructions*.

Runtime on master (1402ms):
![image](https://github.com/user-attachments/assets/cd751048-4eef-40a2-ba28-15f9dfbf2768)

Runtime after optimizations (478ms):
![image](https://github.com/user-attachments/assets/a5add7b3-c1cd-4894-aa3d-d1d36f011fa6)


### Less improvements for token operations
Token operations don't have the sheer number of instructions of a sha256(2048), so they don't benefit as drastically from these speedups. Below are the improvements for combined token mint/transfer/burn/check-balance.

*THIS MIGHT MEAN THAT THE IMPROVEMENTS WON'T BE NOTICEABLE FOR DEVNET/TESTNET*

Runtime on master (736ms)
![image](https://github.com/user-attachments/assets/2f66dfd8-64a3-476b-83ca-516498a368b8)

Runtime after optimizations (537ms)
![image](https://github.com/user-attachments/assets/13f9723e-81fb-45c9-b9ab-fe63f3cfc756)


## Details
1. Cache deserialized instructions by PC 
2. Comment out `logger.trace()`s happening for every memory get/set
3. Simplify tracking of multiple gas dimensions to use hardcoded l2gas and dagas to stop doing *string formatting and reflection* on every gas operation
4. Remove `MeteredTaggedMemory` which would only be needed if we want to charge gas per memory access
5. Change `checkTagsAreSame` to operate on exactly 2 offsets instead of N since it is only ever used on 2
6. Use a static Map to get a `MemoryValue`'s tag from it's constructor name instead of a long if-else chain.
7. Use a static Set to check whether a tag is valid or integral
8. During address resolution, don't keep re-validating and re-getting base memory offset 0 for each loop iteration 
9. Misc optimizations like combining adjacent `assert`s